% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytofkit_GUI.R
\name{chooseFiles_GUI}
\alias{chooseFiles_GUI}
\title{GUI for file selection selection (for cytofast analysis)}
\usage{
chooseFiles_GUI(
  default = getwd(),
  caption = "Choose a file",
  multi = FALSE,
  filters = NULL,
  text
)
}
\arguments{
\item{default}{directory to start with; defult to working directory}

\item{caption}{short to recap the aim of file selection; default to "Choose a file"}

\item{multi}{logical, whether one (FALSE) or multiple (TRUE) could be selected; default FALSE}

\item{filters}{matrix of file extensiosn; see tk_choose.files}

\item{text}{a long text to describe the context of the file selection}
}
\value{
The path of the directory containing FCS files
}
\description{
Returns the selected files; mainly a wrapper of tk_choose.files()
}
\examples{
#chooseDir_GUI()
}
