% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_2RDR.R
\name{create_and_list_folders}
\alias{create_and_list_folders}
\title{Create and List Folders for Project Execution}
\usage{
create_and_list_folders(destinationfile, numero_exe, PROJET)
}
\arguments{
\item{destinationfile}{The base directory where the folders will be created.}

\item{numero_exe}{The number of the execution or experiment (used to create unique folder names).}

\item{PROJET}{The name of the project (used as part of the folder names).}
}
\value{
A list containing project names (FDR, SDR, MAP) and the paths of the created folders.
}
\description{
This function creates a set of folders for organizing output data related to the project, 
including folders for first dimensional reduction (FDR), second dimensional reduction (SDR), and mapping (MAP).
If the folders already exist, it notifies the user. The function also lists the created folders
and returns their paths along with the corresponding project names.
}
\examples{
create_and_list_folders("base_directory", 1, "MyProject")
}
