% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_postProcess.R
\name{cytof_addToFCS}
\alias{cytof_addToFCS}
\title{Add data to the original FCS files}
\usage{
cytof_addToFCS(
  to_add,
  rawFCSdir,
  origSampNames = NULL,
  analyzedFCSdir,
  transformed_cols = c("tsne_1", "tsne_2"),
  cluster_cols = c("cluster")
)
}
\arguments{
\item{to_add}{The data matrix to add/append as columns/channels.}

\item{rawFCSdir}{The directory containing the original fcs files.}

\item{origSampNames}{Vector of original names of samples, if samples were renamed.}

\item{analyzedFCSdir}{The directory to store the new fcs files.}

\item{transformed_cols}{The column name of the dimension transformed data in \code{data}.}

\item{cluster_cols}{The column name of the cluster data in \code{data}.}
}
\value{
New fcs files stored under \code{analyzedFCSdir}
}
\description{
Store the new dimension transformed data and cluster data into the exprs 
matrix in new fcs files under \code{analyzedFCSdir}
}
