% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_helpers.R
\name{cytof_chooseFile}
\alias{cytof_chooseFile}
\title{Open a dialog to select a single file}
\usage{
cytof_chooseFile()
}
\value{
A character vector giving zero or more file paths.
}
\description{
Use the standard file.choose function of the base package .
}
\details{
Unlike file.choose, choose.files will always attempt to return a
  character vector giving a list of files. If the user cancels the dialog,
  then zero files are returned, whereas file.choose would signal an error.
}
\examples{
\dontrun{
filename <- cytof_chooseFile()
}

}
\seealso{
\code{\link{file.choose}}
}
