% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_preProcess.R
\name{cytof_exprsExtract}
\alias{cytof_exprsExtract}
\title{Extract the expression data from a FCS file with preprocessing}
\usage{
cytof_exprsExtract(
  fcsFile,
  verbose = FALSE,
  comp = FALSE,
  transformMethod = c("autoLgcl", "cytofAsinh", "logicle", "arcsinh", "none"),
  scaleTo = NULL,
  ...,
  rename_rc = TRUE
)
}
\arguments{
\item{fcsFile}{The name of the FCS file.}

\item{verbose}{If \verb{TRUE}, print the message details of FCS loading.}

\item{comp}{If \verb{TRUE}, does compensation  by compensation matrix contained in FCS. Agrument also accepts a compensation matrix to be applied. Otherwise \verb{FALSE}.}

\item{transformMethod}{Data Transformation method, including \code{autoLgcl}, \code{cytofAsinh}, \code{logicle} and \code{arcsinh}, or \code{none} to avoid transformation.}

\item{scaleTo}{Scale the expression to a specified range c(a, b), default is NULL.}

\item{...}{parameters passed to transformation.}

\item{rename_rc}{boolean to rename or not rows and columns of the expresion matrix.}
}
\value{
A transformed expression data matrix, row names added as \code{filename_cellID}, column names added as \code{name<desc>}.
}
\description{
Extract the FCS expresssion data with preprocessing of compensation (for FCM data only)
and transformation. Transformtion methods includes \code{autoLgcl}, \code{cytofAsinh}, 
\code{logicle} (customizable) and \code{arcsinh} (customizable).
}
\examples{
d <- system.file('extdata',package='cytofkitlab')
fcsFile <- list.files(d,pattern='.fcs$',full=TRUE)
transformed <- cytof_exprsExtract(fcsFile)
}
