% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_preProcess.R
\name{cytof_getParamFCS}
\alias{cytof_getParamFCS}
\title{Get the parameters aka channel names formatted for cytofkit, i.e. name and
description}
\usage{
cytof_getParamFCS(fcsFile, cat = FALSE)
}
\arguments{
\item{fcsFile}{FCS file name to read.}

\item{cat}{boolean to cat the result on the R console.}
}
\value{
a vector of parameters for cytofkit analysis
}
\description{
Get the parameters aka channel names formatted for cytofkit, i.e. name and
description
}
\examples{
# None yet
}
