% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_helpers.R
\name{cytof_logResults}
\alias{cytof_logResults}
\title{Log information to the cytofkit analysis results}
\usage{
cytof_logResults(messages, time_prefix = TRUE, env = environment())
}
\arguments{
\item{messages}{A vector of strings.}

\item{time_prefix}{If TURE (default), the first strings is prefixed with
Sys.time() function.}

\item{env}{The environment where to look for the analysis results objects
(analysis_results). If missing the current environment is used.}
}
\value{
None.
}
\description{
Append messages to the log of the  analysis results from cytofkit (RData
object). Print the log information if message is missing.
}
\examples{
\dontrun{
cytof_logResults()  # Print log if any
cytof_logResults("UMAP at 50\%", prefix_time = FALSE)
}

}
\seealso{
\code{\link{cytof_writeResults}}
}
