% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_2RDR.R
\name{cytof_map_knn}
\alias{cytof_map_knn}
\title{Map RAW FCS files to cytofkit result and transfer reduced dimensions and
clusterings.}
\usage{
cytof_map_knn(
  ck_res,
  res_map_fcs_dir = "\%s_map_k\%d",
  knn = 1,
  average_dr = "arithmetic",
  prefix = "",
  suffix = "_k\%d",
  comp = TRUE,
  transformMethod,
  ...
)
}
\arguments{
\item{ck_res}{file name of a result}

\item{res_map_fcs_dir}{directory to write the mapped FCS files}

\item{knn}{number of nearest neighbors}

\item{average_dr}{method to average reduced dimension coordinates. Either
"arithmetic", either "softmax" weighted, the latter being slower.}

\item{prefix}{prefix to the resulting FCS file name}

\item{suffix}{suffix to the resulting FCS file name}

\item{comp}{apply compensation}

\item{transformMethod}{transformation method}

\item{...}{parameters passed for the transformation}
}
\value{
none
}
\description{
Map RAW FCS files to cytofkit result and transfer reduced dimensions and
clusterings.
}
\examples{
# None
}
\author{
S. Jully
}
