% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_mapping.R
\name{cytof_mapping}
\alias{cytof_mapping}
\title{CyTOF Data Mapping Function}
\usage{
cytof_mapping(
  rawFCSdir,
  fdrFCRdir,
  fdrprojectname,
  resultDir,
  markers = NULL,
  dims = 2,
  transformMethod = c("autoLgcl", "cytofAsinh", "logicle", "arcsinh", "none"),
  method = c("tsne", "pca", "isomap", "diffusionmap", "umap", "NULL"),
  k = 10,
  algorithm = c("kd_tree", "cover_tree", "CR", "brute")
)
}
\arguments{
\item{rawFCSdir}{The directory containing the raw FCS files.}

\item{fdrFCRdir}{The directory containing the FDR (full-dimensional reduction) analyzed FCS files.}

\item{fdrprojectname}{The project name for the FDR data.}

\item{resultDir}{The directory where the resulting files (mapped FCS and RData) will be saved.}

\item{markers}{A character vector of markers to use for matching between raw and FDR data.}

\item{dims}{The number of dimensions for the pseudo-coordinates (default is 2).}

\item{transformMethod}{The transformation method to apply (options: "autoLgcl", "cytofAsinh", "logicle", "arcsinh", "none").}

\item{method}{The dimensionality reduction method used (options: "tsne", "pca", "isomap", "diffusionmap", "umap", "NULL").}

\item{k}{The number of nearest neighbors to search for in the KNN algorithm.}

\item{algorithm}{The KNN search algorithm (options: "kd_tree", "cover_tree", "CR", "brute").}
}
\value{
A data frame containing the raw data with added pseudo-coordinates.
}
\description{
This function maps raw CyTOF data to a set of pseudo-coordinates derived from dimensionality reduction methods (e.g., UMAP or t-SNE).
It takes raw and processed FCS files as inputs, performs a nearest-neighbor search based on specified markers,
and adds the pseudo-coordinates to the raw data. The function saves the updated raw data with the new coordinates 
as both FCS and RData files and modifies the FCS file descriptions.
}
\examples{
cytof_mapping("path/to/rawFCS", "path/to/fdrFCS", "ProjectName", "path/to/results", markers = c("CD3", "CD19"), dims = 2)
}
