% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_preProcess.R
\name{cytof_readFCS}
\alias{cytof_readFCS}
\title{Read a FCS file with compensation and return a flowFrame}
\usage{
cytof_readFCS(fcsFile, comp = FALSE, verbose = FALSE)
}
\arguments{
\item{fcsFile}{The name of the FCS file.}

\item{comp}{If \verb{TRUE}, does compensation  by compensation matrix
contained in FCS. Argument also accepts a compensation matrix to be
applied. Otherwise \verb{FALSE}.}

\item{verbose}{If \verb{TRUE}, print the message details of FCS loading.}
}
\value{
A flowFrame, read with no transformation at all, except compensation
  if asked. See \code{cytof_transfFCS}, to transform expression.
}
\description{
Read a FCS file with compensation and no transformation.
}
\examples{
d <- system.file('extdata', package = 'cytofkitlab')
fcsFiles <- list.files(d, pattern = '\\\\.fcs$', full=TRUE)
fcs_flowFrame <- cytof_readFCS(fcsFiles[1])
}
