% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_helpers.R
\name{cytof_relocateResults}
\alias{cytof_relocateResults}
\title{Relocate the directory of the input FCS and of the output results}
\usage{
cytof_relocateResults(overwrite = FALSE)
}
\arguments{
\item{overwrite}{If FALSE (default) the modified RData object is stored in
the same directory as the input RData file with a unique prefix.}
}
\value{
The written file name invisibly.
}
\description{
Interactively select the RData file, the directory of the input FCS and of
the output results, modify RData object and store it to disk.
}
\examples{
\dontrun{
cytofkit_RData_file <- cytof_relocateResults()
load(cytofkit_RData_file)
# The RAW FCS files should be at the following location
dir(analysis_results$rawFCSdir)
# If it's OK, save the FCS again
cytof_writeResults(
    analysis_results = analysis_results,
    saveToRData = FALSE,
    saveToFCS = TRUE,
    saveToFiles = FALSE)
dir(analysis_results$resultDir)
}

}
\seealso{
\code{\link{cytof_writeResults}}
}
