% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_preProcess.R
\name{cytof_transfFCS}
\alias{cytof_transfFCS}
\title{Transform the expression/intensity values of a FCS file (flowFrame)}
\usage{
cytof_transfFCS(
  fcs,
  transformMethod = c("autoLgcl", "cytofAsinh", "logicle", "arcsinh", "none"),
  scaleTo = NULL,
  q = 0.05,
  l_w = 0.1,
  l_t = 4000,
  l_m = 4.5,
  l_a = 0,
  a_a = 1,
  a_b = 1,
  a_c = 0,
  rename_rc = TRUE
)
}
\arguments{
\item{fcs}{The flowFrame object after reading a FCS file.}

\item{transformMethod}{Data Transformation method, including \code{autoLgcl},
\code{cytofAsinh}, \code{logicle} and \code{arcsinh}, or \code{none} to
avoid transformation.}

\item{scaleTo}{Scale the expression to a specified range c(a, b), default is
NULL.}

\item{q}{Quantile of negative values removed for auto w estimation, default
is 0.05, parameter for autoLgcl transformation.}

\item{l_w}{Linearization width in asymptotic decades, parameter for logicle
transformation.}

\item{l_t}{Top of the scale data value, parameter for logicle transformation.}

\item{l_m}{Full width of the transformed display in asymptotic decades,
parameter for logicle transformation.}

\item{l_a}{Additional negative range to be included in the display in
asymptotic decades, parameter for logicle transformation.}

\item{a_a}{Positive double that corresponds to the base of the arcsinh
transformation, \code{arcsinh} = asinh(a + b * x) + c.}

\item{a_b}{Positive double that corresponds to a scale factor of the arcsinh
transformation, \code{arcsinh} = asinh(a + b * x) + c.}

\item{a_c}{Positive double that corresponds to another scale factor of the
arcsinh transformation, \code{arcsinh} = asinh(a + b * x) + c.}

\item{rename_rc}{boolean to rename or not rows and columns of the expresion matrix.}
}
\value{
A transformed flowFrame, row names added as \code{filename_cellID},
  column names added as \code{name<desc>}.
}
\description{
Transform the FCS expression data with transformation. Transformation methods
include \code{autoLgcl}, \code{cytofAsinh}, \code{logicle} (customizable) and
\code{arcsinh} (customizable).
}
\examples{
d <- system.file('extdata', package = 'cytofkitlab')
fcsFiles <- list.files(d, pattern = '\\\\.fcs$', full=TRUE)
ff_raw <- cytof_readFCS(fcsFiles[1])
ff_transformed <- cytof_transfFCS(
    ff_raw, transformMethod = "arcsinh", a_a = 0, a_b = 1/5, a_c =0, rename_rc = FALSE)
t(rbind(range(ff_raw, "data"), range(ff_transformed), range(ff_transformed, "data")))
}
