% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cytof_verbose}
\alias{cytof_verbose}
\title{cytof_verbose gets/sets the verbosity level for functions that need it}
\usage{
cytof_verbose(lev)
}
\arguments{
\item{lev}{integer level of verbosity. When defined, this set the level. When
missing the current level is reported. If level has not been defined, 0 is
returned.}
}
\value{
verbosity
}
\description{
cytof_verbose gets/sets the verbosity level for functions that need it
}
\examples{
cytof_verbose()
cytof_verbose(1)
cytof_verbose()
}
