% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytofkit_shiny_explorer.R
\name{cytofkit_shiny_explorer}
\alias{cytofkit_shiny_explorer}
\title{A Shiny Appliaction to interactively explore the analysis results}
\usage{
cytofkit_shiny_explorer(RData = NULL, onServer = FALSE, port = NULL)
}
\arguments{
\item{RData}{Either the RData object file or data object, if missing, RData file need to be
loaded in the Shiny application}

\item{onServer}{Logical value, if \verb{TRUE}, sets shinyApp host to 0.0.0.0 for other clients to
access, otherwise defaults to 127.0.0.1 (local host)}

\item{port}{Integer value, specifies a port for shinyApp for other clients to access, otherwise
automatic selection}
}
\value{
Opens shiny application for results exploration
}
\description{
Take the the RData object file saved by cytofkit as input, automatically load the data and allow
exploration of the analysis results with interactive control
}
\examples{
d <- system.file('extdata',package='cytofkitlab')
Rdata <- list.files(d, pattern = '.RData$', full.names = TRUE)
#only for interactive sessions, remove hash to run
#cytofkitShinyAPP(Rdata)

}
\author{
Samuel Granjeaud
}
