% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytofkitlab-package.R
\docType{package}
\name{cytofkitlabv2-package}
\alias{cytofkitlabv2}
\alias{cytofkitlabv2-package}
\title{cytofkitlabv2: cytofkitlab: a fork of cytofkit, an integrated mass cytometry data analysis pipeline}
\description{
An integrated mass cytometry data analysis pipeline that enables simultaneous illustration of cellular diversity and progression.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/i-cyto/cytofkitlab}
  \item Report bugs at \url{https://github.com/i-cyto/cytofkitlab/issues}
}

}
\author{
\strong{Maintainer}: Samuel Granjeaud \email{samuel.granjeaud@inserm.fr}

Authors:
\itemize{
  \item Jinmiao  Chen \email{Chen_Jinmiao@immunol.a-star.edu.sg}
  \item Hao Chen
  \item Matthew Myint \email{a0124008@u.nus.edu}
}

}
\keyword{internal}
