% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytofkit_GUI.R
\name{fixedAsinhParameters_GUI}
\alias{fixedAsinhParameters_GUI}
\title{GUI for gettting parameter for asinh transformation}
\usage{
fixedAsinhParameters_GUI(fixedAsinhParas = c(0, 1/5, 0))
}
\arguments{
\item{fixedLgclParas}{parameters vector containing w, t, m, a}
}
\value{
Parameters for fixed asinh transformation
}
\description{
Extract the parameter for fixed asinh transformation
}
\examples{
#fixedAsinhParameters_GUI
}
