% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_options}
\alias{merge_options}
\title{Function to merge default options with dots arguments}
\usage{
merge_options(prefix, defaults, dots)
}
\arguments{
\item{prefix}{a character string that identifies the options to retrieve and
merge. The prefix does not contain the separator.}

\item{defaults}{a list of options that hold the default values.}

\item{dots}{a list of options, typically the '...' arguments passed to a
function. Each option/value is named and the names are considered beginning
with prefix followed by the '.' separator.}
}
\value{
a list of named values aka options.
}
\description{
Arguments in the dots list beginning withng the given prefix are retrieved.
Their value replaces the value of the options in the defaults list when the
same name exists in both lists. Otherwise, options of both lists are kept in
the resulting merged list where the prefix is removed
}
\examples{
# None
}
