% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytofast.R
\name{readCytofkitFCS}
\alias{readCytofkitFCS}
\title{Read .fcs files created by Cytofkit and return a cytofast object}
\usage{
readCytofkitFCS(
  dir = NULL,
  colNames = c("description", "names"),
  clusterID = "FlowSOM_clusterIDs",
  cleanUp = TRUE
)
}
\arguments{
\item{dir}{directory containing the .FCS files created by Cytofkit}

\item{colNames}{character string that specifies which label should be used as
the column names. This could be the name of the markers 'names' or the
description for the markers 'description'.}

\item{clusterID}{character string that specifies which label should be used
to retrieve the cluster identifiers. Typically, this should be
"FlowSOM_clusterIDs", "Rphenograph_clusterIDs" or "ClusterX_clusterIDs"}

\item{cleanUp}{logical that specifies to remove extra channels added by
Cytofkit}
}
\value{
The function returns an object of class
  \code{\link[cytofast]{cfList-class}}.
}
\description{
This function reads and combines .FCS files created specifically
  by Cytofokit.
}
\note{
To successfully read the .FCS files created by cytofkit make sure that
  there are no double channels. It is possible to remove the double tag in
  software like Flowjo.

  As for the `colNames` parameter, both 'name' and 'description' will always
  work, it is up to the user to decide which one is most preferable. "" or
  "<NA>" descriptions are replaced by the corresponding name.

  This function is a wrapper around \code{\link[flowCore]{read.FCS}}. For
  more flexibility see their help page. Parameters currently set are:
  transformation=FALSE, truncate_max_range=FALSE, min.limit = NULL

  NB: No compensation nor transformation is applied to the data.
}
\examples{
\dontrun{
# Directory of input FCS
dirFCS <- system.file('extdata',package='cytofkitlab')
# For demo, simulate a dir of results and copy the demo result .RData file
dirRes <- tempdir()
file.copy(file.path(dirFCS, "demo_fcs.RData"), file.path(dirRes, "demo_fcs.RData"))
# Load an annotated result .RData
load(file.path(dirRes, "demo_fcs.RData"))
# Write clustered data with Rphenograph to FCS files in a temporary directory
cytof_writeResults(
    analysis_results,  # from load function
    projectName = "demo",
    saveToRData = FALSE,
    saveToFCS = TRUE,
    saveToFiles = FALSE,
    resultDir = dirRes,
    rawFCSdir = dirFCS,
    inverseLgclTrans = FALSE)
# Verify the FCS files
#   fn <- dir(file.path(tempdir(), paste0("demo","_analyzedFCS")), full.names = TRUE)
#   flowCore::read.FCS(fn)
# Read FCS files into a cytofast cfList object
cfData <- readCytofkitFCS(
    dir = file.path(dirRes, paste0("demo","_analyzedFCS")),
    colNames = "description",
    clusterID = "Rphenograph_clusterIDs"
)
cfData
}

}
\keyword{FCS,}
\keyword{cytofast}
\keyword{data,}
\keyword{read,}
