% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytofkit_GUI.R
\name{storeMarkers_GUI}
\alias{storeMarkers_GUI}
\title{GUI for marker selection}
\usage{
storeMarkers_GUI(markerFile, rawFCSdir, fcsFile)
}
\arguments{
\item{markerFile}{The name of the FCS file}

\item{rawFCSdir}{The path of the FCS file}

\item{fcsFile}{The name of the FCS file}
}
\value{
List of markers for ddimension reduction and clustering
}
\description{
Extract the markers from the fcsfiles, write the selection into a file and returns the directory
of FCS files
}
\examples{
#storeMarkers_GUI()
}
