% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_helpers.R
\name{time_difference}
\alias{time_difference}
\title{Time Difference Calculation}
\usage{
time_difference(start_time, end_time)
}
\arguments{
\item{start_time}{Starting time (POSIXct object).}

\item{end_time}{Ending time (POSIXct object).}
}
\value{
A character string representing the time difference formatted as "HH:MM:SS.SS".
}
\description{
This function calculates the difference between two time points (`start_time` and `end_time`) 
and returns the result formatted as "HH:MM:SS.SS". The difference is measured in seconds.
}
\examples{
start_time <- Sys.time()
Sys.sleep(5)
end_time <- Sys.time()
time_difference(start_time, end_time)  # "00:00:05.00"
}
\author{
S. Jully
}
