% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_time_tracker.R
\name{time_info}
\alias{time_info}
\title{Generate Execution Time Information File}
\usage{
time_info(
  experience_folder,
  numero_exe,
  parameters_FDR,
  parameters_SDR,
  method_dim_reduction_FDR,
  fixedNum_FDR,
  time_dimred_fdr,
  time_cluster_fdr,
  max_nb_knn,
  time_map,
  method_dim_reduction_SDR,
  fixedNum_SDR,
  time_dimred_sdr,
  time_cluster_sdr
)
}
\arguments{
\item{experience_folder}{Directory where the results are stored.}

\item{numero_exe}{Execution number or identifier for this run.}

\item{parameters_FDR}{Parameters used for the full-dimensional reduction (FDR).}

\item{parameters_SDR}{Parameters used for the sub-dimensional reduction (SDR).}

\item{method_dim_reduction_FDR}{Method used for full-dimensional reduction (e.g., t-SNE, UMAP).}

\item{fixedNum_FDR}{Number of cells used for the FDR step.}

\item{time_dimred_fdr}{Execution time for the full-dimensional reduction step (in seconds).}

\item{time_cluster_fdr}{Execution time for the FDR clustering step (in seconds).}

\item{max_nb_knn}{Maximum number of nearest neighbors to search during the mapping phase.}

\item{time_map}{Execution time for the mapping phase (in seconds).}

\item{method_dim_reduction_SDR}{Method used for sub-dimensional reduction (e.g., t-SNE, UMAP).}

\item{fixedNum_SDR}{Number of cells used for the SDR step.}

\item{time_dimred_sdr}{Execution time for the sub-dimensional reduction step (in seconds).}

\item{time_cluster_sdr}{Execution time for the SDR clustering step (in seconds).}
}
\value{
No return value. Writes the information to a text file.
}
\description{
This function records various time information from different stages of the analysis pipeline 
(such as dimensionality reduction, clustering, and mapping) into a text file.
}
\examples{
time_info("results_folder", 1, c("param1", "param2"), c("paramA", "paramB"), 
          "UMAP", 1000, 120.5, 60.3, 50, 180, "t-SNE", 800, 100.1, 45.2)
}
